refactor: Phase 3.1d — extract VisualizationUpdateMixin from ElectrodeAdvisorWidget
Move 6 dynamic 3D visualization methods (904 lines) into
VisualizationUpdateMixin in main_window_visualization_update.py:
- _draw_3d_real_geometry (153 lines)
- _update_3d_visualization (186 lines)
- _draw_3d_conductive_paths_new (118 lines)
- _draw_correct_trapezoidal_path (208 lines)
- _draw_correct_via_metal_path (126 lines)
- _draw_electrode_length_extrusion (113 lines)

Monolith now at 2,249 lines (-48.7% from original 4,387).

5 mixins extracted total:
- DrawingMixin: 510 lines (static geometry)
- InputPanelMixin: 402 lines (input controls)
- VisualControlsMixin: 485 lines (visual controls panel)
- VisualizationUpdateMixin: 934 lines (dynamic 3D updates)
